improve javascript data handling
This commit improves the oath parser output format to be somewhat more
robust, by converting the suit array into a dict. This also updates the
suit count rendering to generate the entire table automatically in a
non-fragile way.

To customize the table order, change the order in the 'suits' array.